quiz&voting&survey img upload added

+++ src/main/java/com/example/notfound_backend/data/repository/utility/UploadedImageRepository.java
@@ -3,5 +3,5 @@ package com.example.notfound_backend.data.repository.utility;
 import com.example.notfound_backend.data.entity.utility.UploadImageEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UploadedImageRepository extends JpaRepository<UploadImageEntity, Long> {
+public interface UploadedImageRepository extends JpaRepository<UploadImageEntity, Integer> {
 }
